Allow draw also for playoff matches
Especially with younger players there can be matches after the group stage so that no final results are calculated. This means that
there's no necessarily need to resolve the winner of the match.

If draw is saved for a normal playoff match, it means that no further matches will be set until the official changes to a winning
socre.